the binary suffix for Debian binaries is set here

+++ prep_deb_package.sh
@@ -1,15 +1,15 @@
 #!/bin/bash
 
 # generate configure, configure
-./autogen.sh
+# the Debian package always has the same prefix
+# export FCFLAGS="-O4 -march=core2 -ftree-vectorize -ffast-math -funroll-loops -finline-functions -fopenmp"
+CRT_PREFIX="dev" ./autogen.sh
 
 # create archive
 make dist-bzip2
 
 # delete if necessary
-if [ -d package ]; then
-	rm -r package
-fi
+test -d package && rm -r package
 
 mkdir package
 
@@ -17,4 +17,3 @@ mv crtomomod*.bz2 package
 cd package
 tar xvjf crtomomod*.bz2
 echo "Now enter package/crtomomod*/ and type dpkg-buildpackage"
-
